requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.36.0
diffusers>=0.25.0
accelerate>=0.25.0
safetensors>=0.4.0
Pillow>=10.0.0
pyyaml>=6.0
bitsandbytes>=0.43.0
sentencepiece>=0.1.99
protobuf>=3.20.0
huggingface_hub>=0.20.0
